TAYLOR: Fix metacommand parsing
Commands like TRANSCRIPT OFF was broken in games that defined OFF
in their internal dictionary (and TRANSCRIPT ON if they defined ON, and so on.)

+++ terps/taylor/taylor.h
@@ -302,7 +302,7 @@ extern size_t AnimationData;
 
 extern size_t VerbBase;
 extern char LastChar;
-extern uint8_t Word[];
+extern uint16_t Word[];
 extern int PendSpace;
 extern int FirstAfterInput;
 extern int LastVerb;
